test: 隔离 macOS XCTest 应用宿主
